backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
sqlalchemy==2.0.35
alembic==1.13.0
psycopg2-binary==2.9.9
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic-settings==2.5.2
langchain==0.3.7
langchain-community==0.3.5
langchain-groq==0.2.1
langchain-huggingface==0.1.2
langchain-pinecone==0.2.0
pinecone==5.4.2
pypdf==4.3.1
sentence-transformers==3.3.0
pydantic==2.9.0
httpx==0.27.0
groq==0.11.0
